pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "clawnet"
version = "3.0.0"
description = "AI-Powered Interactive Security Terminal"
requires-python = ">=3.10"
dependencies = [
    "psutil>=5.9",
    "rich>=13.0",
    "openai>=1.0",
    "python-telegram-bot>=20.0",
    "send2trash>=1.8",
]

[project.optional-dependencies]
memory = [
    "supermemory>=3.0",
]

[project.scripts]
clawnet = "clawnet:main"

[tool.setuptools]
py-modules = ["clawnet"]
packages   = ["core"]
